Move fetching tracks to its own crate
